+++ packages/blade/src/components/Input/TextInput/TextInput.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState, useRef } from 'react';
 import type { ReactElement, ReactNode } from 'react';
 import type { TextInput as TextInputReactNative } from 'react-native';
 import type { BaseInputProps } from '../BaseInput';
@@ -21,9 +21,13 @@ import { useMergeRefs } from '~utils/useMergeRefs';
 import type {
   BladeElementRef,
   BladeElementRefWithValue,
+  ContainerElementType,
   DataAnalyticsAttribute,
 } from '~utils/types';
 import { hintMarginTop } from '~components/Form/formTokens';
+import { Divider } from '~components/Divider';
+import { getComponentId } from '~utils/isValidAllowedChildren';
+import { DropdownOverlay } from '~components/Dropdown';
 
 // Users should use PasswordInput for input type password
 type Type = Exclude<BaseInputProps['type'], 'password'>;
@@ -81,7 +85,7 @@ type TextInputCommonProps = Pick<
 
   /**
    * Icon that will be rendered at the beginning of the input field
-   * @deprecated Use `leadingIcon` instead. This prop will be removed in the next major version.
+   * @deprecated Use `leading` instead. This prop will be removed in the next major version.
    */
   icon?: IconComponent;
   /**
@@ -98,6 +102,15 @@ type TextInputCommonProps = Pick<
    * @default text
    */
   type?: Type;
+  /**
+   *
+   * Icon or React Element to be rendered at the end of the input field
+   */
+  trailing?: React.ReactElement | IconComponent;
+  /**
+   * Icon or React Element to be rendered at the beginning of the input field
+   */
+  leading?: React.ReactElement | IconComponent;
 } & TaggedInputProps &
   StyledPropsBlade;
 
@@ -178,6 +191,8 @@ const _TextInput: React.ForwardRefRenderFunction<BladeElementRef, TextInputProps
     isTaggedInput,
     tags,
     onTagChange,
+    trailing,
+    leading,
     ...rest
   },
   ref,
@@ -203,52 +218,155 @@ const _TextInput: React.ForwardRefRenderFunction<BladeElementRef, TextInputProps
     value,
     inputRef: textInputRef,
   });
+  const [isTrailingDropDownOpen, setIsTrailingDropDownOpen] = React.useState(false);
+  const [isLeadingDropDownOpen, setIsLeadingDropDownOpen] = React.useState(false);
+  const textInputWrapperRef = useRef<ContainerElementType | null>(null);
+
+  useEffect(() => {
+    if (isTrailingDropDownOpen && isLeadingDropDownOpen) {
+      setIsLeadingDropDownOpen(false);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isTrailingDropDownOpen]);
+
+  useEffect(() => {
+    if (isLeadingDropDownOpen && isTrailingDropDownOpen) {
+      setIsTrailingDropDownOpen(false);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isLeadingDropDownOpen]);
+
+  const leadingDropDown =
+    leading && getComponentId(leading as React.ReactElement) === 'Dropdown' ? leading : null;
+
+  const trailingDropdown =
+    trailing && getComponentId(trailing as React.ReactElement) === 'Dropdown' ? trailing : null;
+  // we need to look into name of component and check if it 's and icon or a dropdown
+  const _leadingIcon: IconComponent | undefined =
+    leading && typeof leading === 'function' && leading.name?.endsWith('Icon')
+      ? (leading as IconComponent)
+      : undefined;
+
+  const _trailingIcon: IconComponent | undefined =
+    trailing && typeof trailing === 'function' && trailing.name?.endsWith('Icon')
+      ? (trailing as IconComponent)
+      : undefined;
+  const hasLeadingInteractionElement = !_leadingIcon && !leadingDropDown && leading;
+
+  const hasTrailingInteractionElement = !_trailingIcon && !trailingDropdown && trailing;
+
+  const renderDropdown = (
+    dropdown: React.ReactElement,
+    isOpen: boolean,
+    setIsOpen: (isOpen: boolean) => void,
+    defaultPlacement: 'bottom-start' | 'bottom-end',
+  ): React.ReactElement | null => {
+    if (!dropdown) {
+      return null;
+    }
+    return React.cloneElement(dropdown, {
+      selectionType: 'single',
+      isOpen,
+      onOpenChange: (isOpen: boolean) => {
+        setIsOpen(isOpen);
+      },
+      children: React.Children.map(dropdown.props.children, (child) => {
+        if (child.type === DropdownOverlay) {
+          return React.cloneElement(child, {
+            referenceRef: textInputWrapperRef,
+            _isNestedDropdown: true,
+            defaultPlacement,
+          });
+        }
+        return child;
+      }),
+    });
+  };
+
+  const renderLeadingDropDown = renderDropdown(
+    leadingDropDown as React.ReactElement,
+    isLeadingDropDownOpen,
+    setIsLeadingDropDownOpen,
+    'bottom-start',
+  );
+  const renderTrailingDropDown = renderDropdown(
+    trailingDropdown as React.ReactElement,
+    isTrailingDropDownOpen,
+    setIsTrailingDropDownOpen,
+    'bottom-end',
+  );
 
   React.useEffect(() => {
     setShouldShowClearButton(Boolean(showClearButton && (defaultValue ?? value)));
   }, [showClearButton, defaultValue, value]);
+
+  const renderClearButton = (): React.ReactElement => {
+    return (
+      <IconButton
+        size="medium"
+        icon={CloseIcon}
+        onClick={() => {
+          if (isEmpty(value) && textInputRef.current) {
+            // when the input field is uncontrolled take the ref and clear the input and then call the onClearButtonClick function
+            if (isReactNative(textInputRef.current)) {
+              textInputRef.current.clear();
+              textInputRef.current.focus();
+            } else if (textInputRef.current instanceof HTMLInputElement) {
+              textInputRef.current.value = '';
+              textInputRef.current.focus();
+            }
+          }
+          handleTagsClear();
+          // if the input field is controlled just call the click handler and the value change shall be left upto the consumer
+          onClearButtonClick?.();
+          textInputRef?.current?.focus();
+          setShouldShowClearButton(false);
+        }}
+        isDisabled={isDisabled}
+        accessibilityLabel="Clear Input Content"
+      />
+    );
+  };
+  const hasTrailingDropDown = Boolean(trailingDropdown);
 
   const renderInteractionElement = (): ReactNode => {
     if (isLoading) {
       return <Spinner accessibilityLabel="Loading Content" color="primary" />;
     }
 
-    if (shouldShowClearButton) {
+    if (shouldShowClearButton && hasTrailingDropDown) {
       return (
-        <IconButton
-          size="medium"
-          icon={CloseIcon}
-          onClick={() => {
-            if (isEmpty(value) && textInputRef.current) {
-              // when the input field is uncontrolled take the ref and clear the input and then call the onClearButtonClick function
-              if (isReactNative(textInputRef.current)) {
-                textInputRef.current.clear();
-                textInputRef.current.focus();
-              } else if (textInputRef.current instanceof HTMLInputElement) {
-                textInputRef.current.value = '';
-                textInputRef.current.focus();
-              }
-            }
-            handleTagsClear();
-            // if the input field is controlled just call the click handler and the value change shall be left upto the consumer
-            onClearButtonClick?.();
-            textInputRef?.current?.focus();
-            setShouldShowClearButton(false);
-          }}
-          isDisabled={isDisabled}
-          accessibilityLabel="Clear Input Content"
-        />
+        <BaseBox display="flex" gap="spacing.3">
+          {renderClearButton()} <Divider orientation="vertical" />
+        </BaseBox>
       );
     }
 
+    if (showClearButton && hasTrailingInteractionElement) {
+      return (
+        <BaseBox display="flex" gap="spacing.3">
+          {renderClearButton()} <Divider orientation="vertical" /> {trailing as React.ReactElement}
+        </BaseBox>
+      );
+    }
+
+    if (shouldShowClearButton) {
+      return renderClearButton();
+    }
+
+    if (hasTrailingInteractionElement) {
+      return trailing as React.ReactElement;
+    }
     return null;
   };
-
   return (
     <BaseInput
       id="textinput"
       componentName={MetaConstants.TextInput}
       ref={mergedRef}
+      setInputWrapperRef={(wrapperNode) => {
+        textInputWrapperRef.current = wrapperNode;
+      }}
       label={label as string}
       accessibilityLabel={accessibilityLabel}
       hideLabelText={!Boolean(label)}
@@ -264,6 +382,11 @@ const _TextInput: React.ForwardRefRenderFunction<BladeElementRef, TextInputProps
       maxTagRows="single"
       activeTagIndex={activeTagIndex}
       setActiveTagIndex={setActiveTagIndex}
+      leadingDropDown={renderLeadingDropDown}
+      trailingDropDown={renderTrailingDropDown}
+      leadingInteractionElement={
+        hasLeadingInteractionElement ? (leading as React.ReactElement) : null
+      }
       onChange={({ name, value }) => {
         if (showClearButton && value?.length) {
           // show the clear button when the user starts typing in
@@ -294,10 +417,10 @@ const _TextInput: React.ForwardRefRenderFunction<BladeElementRef, TextInputProps
       isDisabled={isDisabled}
       necessityIndicator={necessityIndicator}
       isRequired={isRequired}
-      leadingIcon={leadingIcon ?? icon}
+      leadingIcon={_leadingIcon ?? leadingIcon ?? icon}
       prefix={prefix}
       trailingInteractionElement={renderInteractionElement()}
-      trailingIcon={trailingIcon}
+      trailingIcon={_trailingIcon ?? trailingIcon}
       suffix={suffix}
       validationState={validationState}
       errorText={errorText}